# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ZKTeco Relay API
  version: 1.2.0
  description: |
    Windows ZKTeco 脱机通讯 REST/WebSocket 中继服务。
    除健康检查和文档端点外，所有请求均需提供 X-API-Key。
servers:
  - url: http://127.0.0.1:5080
    description: 默认本机服务
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/devices/{deviceId}/users:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    get:
      tags: [Users]
      operationId: listUsers
      summary: 获取人员列表
      responses:
        '200':
          description: 人员列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceUnavailable'
        '502':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/users/{enrollNumber}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/EnrollNumber'
    get:
      tags: [Users]
      operationId: getUser
      summary: 获取人员
      responses:
        '200':
          description: 人员信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
    put:
      tags: [Users]
      operationId: upsertUser
      summary: 新建或更新人员、卡号和密码
      description: password 或 cardNumber 为 null 时保留设备中的已有值。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertUserRequest'
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/InvalidOrVendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
    delete:
      tags: [Users]
      operationId: deleteUser
      summary: 删除人员及登记数据
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/VendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/users/{enrollNumber}/fingerprints/{fingerIndex}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/EnrollNumber'
      - $ref: '#/components/parameters/FingerIndex'
    get:
      tags: [Biometrics]
      operationId: getFingerprint
      summary: 获取指纹模板
      responses:
        '200':
          description: 指纹模板
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FingerprintTemplateResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
    put:
      tags: [Biometrics]
      operationId: setFingerprint
      summary: 写入指纹模板
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FingerprintTemplateRequest'
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/InvalidOrVendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
    delete:
      tags: [Biometrics]
      operationId: deleteFingerprint
      summary: 删除指纹模板
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/VendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/users/{enrollNumber}/face:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/EnrollNumber'
    get:
      tags: [Biometrics]
      operationId: getFace
      summary: 获取人脸模板
      parameters:
        - $ref: '#/components/parameters/FaceIndex'
      responses:
        '200':
          description: 人脸模板
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceTemplateResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
    put:
      tags: [Biometrics]
      operationId: setFace
      summary: 写入人脸模板
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaceTemplateRequest'
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/InvalidOrVendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
    delete:
      tags: [Biometrics]
      operationId: deleteFace
      summary: 删除人脸模板
      parameters:
        - $ref: '#/components/parameters/FaceIndex'
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/VendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/users/{enrollNumber}/photo:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/EnrollNumber'
    get:
      tags: [Biometrics]
      operationId: downloadUserPhoto
      summary: 下载设备原始 JPG 用户照片
      description: 返回设备保存的原始照片，不是人脸算法模板；Relay 不持久化照片。
      responses:
        '200':
          description: 用户 JPG 照片
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPhotoResult'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceUnavailable'
        '422':
          $ref: '#/components/responses/CapabilityNotSupported'
        '502':
          $ref: '#/components/responses/DeviceOperationFailed'
    put:
      tags: [Biometrics]
      operationId: uploadUserPhoto
      summary: 上传人员照片或可见光人脸照片
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPhotoRequest'
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/InvalidOrVendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceUnavailable'
        '422':
          $ref: '#/components/responses/CapabilityNotSupported'
        '502':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/capabilities:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    get:
      tags: [Access]
      operationId: getDeviceCapabilities
      summary: 探测设备门禁和人员 ID 能力
      responses:
        '200':
          description: 分项设备能力；未知项通过 null 和 probeErrors 表示。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceCapabilities'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceUnavailable'
        '502':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/access/door-state:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    get:
      tags: [Access]
      operationId: getDoorState
      summary: 获取当前门状态
      responses:
        '200':
          description: 门状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoorStateResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceUnavailable'
        '502':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/access/normally-open/start:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    post:
      tags: [Access]
      operationId: startNormallyOpen
      summary: 开始常开
      description: 仅适用于 GetACFun=14 的设备，并返回常开前锁驱动时长。
      responses:
        '200':
          description: 常开状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoorModeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceUnavailable'
        '422':
          $ref: '#/components/responses/CapabilityNotSupported'
        '502':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/access/normally-open/end:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    post:
      tags: [Access]
      operationId: endNormallyOpen
      summary: 结束常开并恢复锁驱动时长
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndNormallyOpenRequest'
      responses:
        '200':
          description: 恢复后的门模式
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoorModeResult'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceUnavailable'
        '502':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/access/unlock:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    post:
      tags: [Access]
      operationId: unlockDoor
      summary: 远程开门
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoorUnlockRequest'
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/InvalidOrVendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/access/time-zones/{index}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/TimeZoneIndex'
    get:
      tags: [Access]
      operationId: getAccessTimeZone
      summary: 获取门禁时间段
      responses:
        '200':
          description: 时间段
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeZoneInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
    put:
      tags: [Access]
      operationId: setAccessTimeZone
      summary: 设置门禁时间段
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeZoneRequest'
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/InvalidOrVendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/access/groups/{groupNumber}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/GroupNumber'
    get:
      tags: [Access]
      operationId: getAccessGroup
      summary: 获取门禁权限组
      responses:
        '200':
          description: 权限组
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessGroupInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
    put:
      tags: [Access]
      operationId: setAccessGroup
      summary: 设置门禁权限组
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessGroupRequest'
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/InvalidOrVendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/access/users/{enrollNumber}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/EnrollNumber'
    get:
      tags: [Access]
      operationId: getUserAccess
      summary: 获取人员门禁权限
      responses:
        '200':
          description: 人员门禁权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccessInfo'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
    put:
      tags: [Access]
      operationId: setUserAccess
      summary: 设置人员门禁权限
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAccessRequest'
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/InvalidOrVendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/access/unlock-combinations/{number}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/CombinationNumber'
    get:
      tags: [Access]
      operationId: getUnlockCombination
      summary: 获取多人开锁组合
      responses:
        '200':
          description: 多人开锁组合
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnlockCombinationInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
    put:
      tags: [Access]
      operationId: setUnlockCombination
      summary: 设置多人开锁组合
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnlockCombinationRequest'
      responses:
        '200':
          $ref: '#/components/responses/OperationSucceeded'
        '400':
          $ref: '#/components/responses/InvalidOrVendorFailure'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceOperationFailed'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    ApiKeyQuery:
      type: apiKey
      in: query
      name: apiKey
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      description: 调用方定义的稳定设备标识，最长 64 个字符。
      schema:
        type: string
        maxLength: 64
      example: office-gate-01
    EnrollNumber:
      name: enrollNumber
      in: path
      required: true
      description: 设备中的人员工号。
      schema:
        type: string
      example: '10001'
    FingerIndex:
      name: fingerIndex
      in: path
      required: true
      schema:
        type: integer
        minimum: 0
        maximum: 9
    FaceIndex:
      name: faceIndex
      in: query
      required: false
      schema:
        type: integer
        default: 50
    TimeZoneIndex:
      name: index
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    GroupNumber:
      name: groupNumber
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  responses:
    Unauthorized:
      description: 缺少或提供了错误的 API Key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InvalidRequest:
      description: 请求参数无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InvalidOrVendorFailure:
      description: 请求参数无效或厂商 SDK 拒绝操作
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ApiError'
              - $ref: '#/components/schemas/OperationResult'
    VendorFailure:
      description: 厂商 SDK 拒绝操作
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OperationResult'
    DeviceNotFound:
      description: 设备会话不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: device_not_found
            message: Device was not found.
            vendorErrorCode: null
...
    DeviceUnavailable:
      description: 设备会话存在但尚未连接
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    CapabilityNotSupported:
      description: 当前固件或 SDK 不支持该能力
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    DeviceOperationFailed:
      description: 厂商 SDK 操作失败或返回无效数据
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    OperationSucceeded:
      description: 设备操作成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OperationResult'
  schemas:
...
    UserInfo:
      type: object
      required: [enrollNumber, name, privilege, enabled, cardNumber, hasPassword]
      properties:
        enrollNumber:
          type: string
        name:
          type: string
        privilege:
          type: integer
        enabled:
          type: boolean
        cardNumber:
          type: string
        hasPassword:
          type: boolean
    UpsertUserRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        password:
          type: string
          nullable: true
        privilege:
          type: integer
          default: 0
        enabled:
          type: boolean
          default: true
        cardNumber:
          type: string
          nullable: true
    FingerprintTemplateRequest:
      type: object
      required: [templateData]
      properties:
        fingerIndex:
          type: integer
          description: 路径参数会覆盖该值。
        templateData:
          type: string
    FingerprintTemplateResult:
      type: object
      required: [enrollNumber, fingerIndex, templateData, templateLength]
      properties:
        enrollNumber:
          type: string
        fingerIndex:
          type: integer
        templateData:
          type: string
        templateLength:
          type: integer
    FaceTemplateRequest:
      type: object
      required: [templateData]
      properties:
        templateData:
          type: string
        faceIndex:
          type: integer
          default: 50
    FaceTemplateResult:
      type: object
      required: [enrollNumber, faceIndex, templateData, templateLength]
      properties:
        enrollNumber:
          type: string
        faceIndex:
          type: integer
        templateData:
          type: string
        templateLength:
          type: integer
    UserPhotoRequest:
      type: object
      required: [base64Jpeg]
      properties:
        base64Jpeg:
          type: string
          format: byte
        visibleLightFacePhoto:
          type: boolean
          default: false
    UserPhotoResult:
      type: object
      required: [enrollNumber, fileName, base64Jpeg, byteLength]
      properties:
        enrollNumber:
          type: string
        fileName:
          type: string
          pattern: '^[A-Za-z0-9_-]+\.jpg$'
        base64Jpeg:
          type: string
          format: byte
        byteLength:
          type: integer
          minimum: 4
          maximum: 10485760
    DeviceCapabilities:
      type: object
      required:
        - supportsAdvancedAccess
        - supportsNormallyOpen
        - supportsDoorState
        - supportsUserPhotoDownload
        - supportsAttendanceRangeQuery
        - probeErrors
      properties:
        pinWidth:
          type: integer
          nullable: true
        supportsAlphabeticPin:
          type: boolean
          nullable: true
        accessControlFunction:
          type: integer
          nullable: true
          description: GetACFun 原始值；14 表示高级门禁和常开。
        supportsAdvancedAccess:
          type: boolean
        supportsNormallyOpen:
          type: boolean
        supportsDoorState:
          type: boolean
        supportsUserPhotoDownload:
          type: boolean
          nullable: true
          description: null 表示当前 SDK 无法完成 IsNewFirmwareMachine 探测。
        supportsAttendanceRangeQuery:
          type: boolean
        probeErrors:
          type: array
          items:
            type: string
    DoorStateResult:
      type: object
      required: [open, rawState]
      properties:
        open:
          type: boolean
        rawState:
          type: integer
    EndNormallyOpenRequest:
      type: object
      required: [restoreLockDriveTime]
      properties:
        restoreLockDriveTime:
          type: integer
          minimum: 0
          maximum: 254
    DoorModeResult:
      type: object
      required: [normallyOpen, lockDriveTime]
      properties:
        normallyOpen:
          type: boolean
        lockDriveTime:
          type: integer
          minimum: 0
          maximum: 255
        previousLockDriveTime:
          type: integer
          nullable: true
        doorOpen:
          type: boolean
          nullable: true
    DoorUnlockRequest:
      type: object
      properties:
        delayTenthsOfSecond:
          type: integer
          default: 30
          description: 开门延时，单位为 0.1 秒。
    TimeZoneRequest:
      type: object
      required: [schedule]
      properties:
        timeZoneIndex:
          type: integer
          description: 路径参数会覆盖该值。
        schedule:
          type: string
          minLength: 56
          maxLength: 56
          pattern: '^[0-9]{56}$'
    TimeZoneInfo:
      type: object
      required: [timeZoneIndex, schedule]
      properties:
        timeZoneIndex:
          type: integer
        schedule:
          type: string
...
    AccessGroupInfo:
      allOf:
        - $ref: '#/components/schemas/AccessGroupRequest'
        - type: object
          required: [groupNumber]
...